First pass at moving most Pure functionality into a base class.
