+++ pure_adaptor/pure/__init__.py
@@ -1,6 +1,6 @@
 import logging
 from collections import namedtuple
-from . import v59, v512
+from . import v59
 
 logger = logging.getLogger(__name__)
 
@@ -8,8 +8,7 @@ PureInterface = namedtuple(
     'PureInterface',
     [
         'API',
-        'Dataset',
-        'DownloadManager'
+        'Dataset'
     ]
 )
 
@@ -17,12 +16,6 @@ API_VERSION_MAPPINGS = {
     'v59': PureInterface(
         API=v59.PureAPI,
         Dataset=v59.PureDataset,
-        DownloadManager=v59.PureDownloadManager,
-    ),
-    'v512': PureInterface(
-        API=v512.PureAPI,
-        Dataset=v512.PureDataset,
-        DownloadManager=v512.PureDownloadManager,
     )
 }
