+++ pure_adaptor/pure/base/base_mapping.txt
@@ -1,0 +1,1 @@
+{}
