+++ pure_adaptor/pure/base/checksum.py
