+++ pure_adaptor/pure/base/tests/test_checksum.py
