--- pure_adaptor/pure/v512/__init__.py
@@ -1,5 +1,0 @@
-from .api import PureAPI
-from .models import PureDataset
-from .download_manager import PureDownloadManager
-
-__all__ = ['PureAPI', 'PureDataset', 'PureDownloadManager']
