--- pure_adaptor/pure/v512/research_object_mapping.txt
@@ -1,107 +1,0 @@
-{
-  objectUuid: uuid,
-  objectTitle: title.value || join(`, `, title[].value),
-  objectPersonRole: personAssociations[*].{
-      person: {
-        personUuid: person.uuid || externalPerson.uuid || uuid4(),
-        personGivenNames: name.firstName,
-        personFamilyNames: name.lastName,
-        personOrganisationUnit: {
-            organisationUnitUuid: organisationalUnits[0].uuid || uuid4(),
-            organisationUnitName: organisationalUnits[0].name[0].value || 'not present',
-            organisation: {
-                organisationJiscId : jisc_id(@),
-                organisationName: 'not present',
-                organisationType: `8`,
-                organisationAddress: org_addr(@)
-            }
-        },
-        personIdentifier: [{
-            personIdentifierValue: to_string(id),
-            personIdentifierType: `1`
-        }]
-        },
-      role: personRole[0].value | person_role(@)
-    } || `[]`,
-  objectDescription: description[0].value || 'not present',
-  objectRights: {
-    licence: [{
-                licenceName: documents[0].documentLicense[0].value || 'not present',
-                licenceIdentifier : documents[0].documentLicense[0].uri || 'not present'
-                }],
-    access: [{
-                accessType: `1`,
-                accessStatement: 'not present'
-                }]
-  },
-  objectDate: info | object_date(@),
-  objectKeywords: keywordGroups[].keywords[].value || `[]`,
-  objectCategory: `[]`,
-  objectResourceType: type[0].value | object_resourcetype(@),
-  objectValue: `1`,
-  objectIdentifier: doi | [{
-        identifierValue: @,
-        identifierType: `4`
-  }],
-  objectRelatedIdentifier:relatedProjects[*].{
-    identifier: {
-     identifierValue: uuid,
-     identifierType: `16`
-    },
-    relationType: `1`
-  } || `[]`,
-  objectOrganisationRole: [{
-    organisation: {
-        organisationJiscId: publisher.name[0].value | jisc_id(@),
-        organisationName: publisher.name[0].value,
-        organisationType: `8`,
-        organisationAddress: org_addr(@)
-    },
-    role: publisher.type[0].value | org_type(@)
-  }],
-  objectPreservationEvent: `[]`,
-  objectFile: documents[*].{
-    fileUuid: uuid4(),
-    fileIdentifier: url,
-    fileName: title,
-    fileSize: '',
-    fileLabel: 'not present',
-    fileDateCreated: created,
-    fileRights: @.{
-      rightsStatement: ['not present'],
-      rightsHolder: ['not present']
-      licence: documentLicense[*].{
-        licenceName: value || 'not present',
-        licenceIdentifier: uri || 'not present'
-      } || {
-            rightsStatement: ['not present'],
-            rightsHolder: ['not present'],
-            licence: [{
-                licenceName: 'not present',
-                licenceIdentifier : 'not present'
-            }]
-        },
-      access: visibility[*].{
-        accessType: `1`,
-        accessStatement: 'not present'
-      }
-    },
-    fileChecksum: '',
-    fileFormatType: 'not present',
-    fileCompositionLevel: 'not present',
-    fileUse: `1`,
-    filePreservationEvent: [{
-        preservationEventValue: 'not present',
-        preservationEventType: `3`}],
-    fileStoragePlatform: {
-        storagePlatformUuid: uuid4(),
-        storagePlatformName: 'not present',
-        storagePlatformType: `1`,
-        storagePlatformCost: 'not present'
-        }
-    fileLastDownloaded: created,
-    fileDateModified : [created],
-    fileStorageLocation: 'na',
-    fileTechnicalAttributes: ['not present']
-  } || `[]`
-}
