+++ pure_adaptor/pure/v59/__init__.py
@@ -1,5 +1,4 @@
 from .api import PureAPI
 from .models import PureDataset
-from .download_manager import PureDownloadManager
 
-__all__ = ['PureAPI', 'PureDataset', 'PureDownloadManager']
+__all__ = ['PureAPI', 'PureDataset']
